Change return type and params in divide method to float

+++ src/main/java/com/example/springdocker/mathcalc/Calculator.java
@@ -17,7 +17,7 @@ public class Calculator {
         return a * b;
     }
 
-    public int divide(int a, int b) {
+    public float divide(float a, float b) {
         if(b==0) {
             throw new ArithmeticException("Can't divide by zero");
         } else {
